requirements.txt
aiofiles>=24.1.0
aiogram>=3.14.0
aiohappyeyeballs>=2.4.3
aiohttp<=3.9.0
aiosignal>=1.3.1
annotated-types>=0.7.0
async-timeout>=4.0.3
attrs>=24.2.0
certifi>=2024.8.30
colorama>=0.4.6
frozenlist>=1.5.0
idna>=3.10
intercept-it>=0.0.1
loguru>=0.7.2
magic-filter>=1.0.12
multidict>=6.1.0
propcache>=0.2.0
pydantic<2.10
pydantic_core<=2.23.4
python-dotenv>=1.0.1
pytz>=2024.2
typing_extensions>=4.12.2
win32-setctime>=1.1.0
yarl>=1.17.1
